fix(refactor): utils dir moved under lib
